+++ packages/core/CHANGELOG.md
@@ -3,6 +3,108 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.3](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr@0.5.2...@tpluscode/fun-ddr@0.5.3) (2020-01-07)
+
+**Note:** Version bump only for package @tpluscode/fun-ddr
+
+
+
+
+
+## [0.5.2](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr@0.5.1...@tpluscode/fun-ddr@0.5.2) (2019-12-06)
+
+
+### Bug Fixes
+
+* type aggregate root to allow saving to repo of base type ([f83f7b1](https://github.com/tpluscode/fun-ddr/commit/f83f7b1))
+
+
+
+
+
+## [0.5.1](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr@0.5.0...@tpluscode/fun-ddr@0.5.1) (2019-12-03)
+
+
+### Bug Fixes
+
+* make event emitting awaitable ([b1dc3e8](https://github.com/tpluscode/fun-ddr/commit/b1dc3e8))
+
+
+
+
+
+# [0.5.0](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr@0.4.0...@tpluscode/fun-ddr@0.5.0) (2019-11-29)
+
+
+### Features
+
+* add a function to directly emit any domain event ([3cae699](https://github.com/tpluscode/fun-ddr/commit/3cae699))
+
+
+
+
+
+# [0.4.0](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr@0.3.1...@tpluscode/fun-ddr@0.4.0) (2019-11-18)
+
+
+### Bug Fixes
+
+* eventemitter should be a dependency ([aa348ee](https://github.com/tpluscode/fun-ddr/commit/aa348ee))
+
+
+### Features
+
+* use debug for logging event handlers ([c616be9](https://github.com/tpluscode/fun-ddr/commit/c616be9))
+
+
+
+
+
+## [0.3.1](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr@0.3.0...@tpluscode/fun-ddr@0.3.1) (2019-10-28)
+
+
+### Bug Fixes
+
+* make error detail fields public ([8a6b580](https://github.com/tpluscode/fun-ddr/commit/8a6b580))
+
+
+
+
+
+# [0.3.0](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr@0.2.0...@tpluscode/fun-ddr@0.3.0) (2019-10-28)
+
+
+### Features
+
+* adding specialised Error types ([b7b3cca](https://github.com/tpluscode/fun-ddr/commit/b7b3cca))
+
+
+
+
+
+# [0.2.0](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr@0.1.1...@tpluscode/fun-ddr@0.2.0) (2019-10-16)
+
+
+### Features
+
+* return a function which can be used to invoke handler ([033d5d4](https://github.com/tpluscode/fun-ddr/commit/033d5d4))
+
+
+
+
+
+## [0.1.1](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr@0.1.0...@tpluscode/fun-ddr@0.1.1) (2019-10-02)
+
+
+### Bug Fixes
+
+* make aggregate root async to wait for modifications to finish ([7634cc9](https://github.com/tpluscode/fun-ddr/commit/7634cc9))
+* throw aggregate error before null state error ([e342878](https://github.com/tpluscode/fun-ddr/commit/e342878))
+
+
+
+
+
 # 0.1.0 (2019-10-01)