+++ packages/sparql-graph-repository/CHANGELOG.md
@@ -3,6 +3,138 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.10](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.2.9...@tpluscode/fun-ddr-sparql-graph-repository@0.2.10) (2020-01-07)
+
+**Note:** Version bump only for package @tpluscode/fun-ddr-sparql-graph-repository
+
+
+
+
+
+## [0.2.9](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.2.8...@tpluscode/fun-ddr-sparql-graph-repository@0.2.9) (2019-12-06)
+
+**Note:** Version bump only for package @tpluscode/fun-ddr-sparql-graph-repository
+
+
+
+
+
+## [0.2.8](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.2.7...@tpluscode/fun-ddr-sparql-graph-repository@0.2.8) (2019-12-03)
+
+**Note:** Version bump only for package @tpluscode/fun-ddr-sparql-graph-repository
+
+
+
+
+
+## [0.2.7](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.2.6...@tpluscode/fun-ddr-sparql-graph-repository@0.2.7) (2019-11-29)
+
+**Note:** Version bump only for package @tpluscode/fun-ddr-sparql-graph-repository
+
+
+
+
+
+## [0.2.6](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.2.5...@tpluscode/fun-ddr-sparql-graph-repository@0.2.6) (2019-11-26)
+
+
+### Bug Fixes
+
+* **stardog:** base URI must be repeated in multiple update queries ([1c9b3ce](https://github.com/tpluscode/fun-ddr/commit/1c9b3ce))
+* **stardog:** remove leading whitespace from queries to prevent errors ([ecb2eca](https://github.com/tpluscode/fun-ddr/commit/ecb2eca))
+
+
+
+
+
+## [0.2.5](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.2.4...@tpluscode/fun-ddr-sparql-graph-repository@0.2.5) (2019-11-26)
+
+
+### Bug Fixes
+
+* throw errors on failed SPARQL queries ([5620e41](https://github.com/tpluscode/fun-ddr/commit/5620e41))
+
+
+
+
+
+## [0.2.4](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.2.3...@tpluscode/fun-ddr-sparql-graph-repository@0.2.4) (2019-11-18)
+
+**Note:** Version bump only for package @tpluscode/fun-ddr-sparql-graph-repository
+
+
+
+
+
+## [0.2.3](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.2.2...@tpluscode/fun-ddr-sparql-graph-repository@0.2.3) (2019-11-18)
+
+**Note:** Version bump only for package @tpluscode/fun-ddr-sparql-graph-repository
+
+
+
+
+
+## [0.2.2](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.2.1...@tpluscode/fun-ddr-sparql-graph-repository@0.2.2) (2019-11-01)
+
+
+### Bug Fixes
+
+* correctly deserialize rdf:List ([0c73709](https://github.com/tpluscode/fun-ddr/commit/0c73709))
+
+
+
+
+
+## [0.2.1](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.2.0...@tpluscode/fun-ddr-sparql-graph-repository@0.2.1) (2019-10-28)
+
+**Note:** Version bump only for package @tpluscode/fun-ddr-sparql-graph-repository
+
+
+
+
+
+# [0.2.0](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.1.5...@tpluscode/fun-ddr-sparql-graph-repository@0.2.0) (2019-10-28)
+
+
+### Features
+
+* adding specialised Error types ([b7b3cca](https://github.com/tpluscode/fun-ddr/commit/b7b3cca))
+
+
+
+
+
+## [0.1.5](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.1.4...@tpluscode/fun-ddr-sparql-graph-repository@0.1.5) (2019-10-16)
+
+**Note:** Version bump only for package @tpluscode/fun-ddr-sparql-graph-repository
+
+
+
+
+
+## [0.1.4](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.1.3...@tpluscode/fun-ddr-sparql-graph-repository@0.1.4) (2019-10-09)
+
+
+### Bug Fixes
+
+* include the id when an aggregate is not found ([0765177](https://github.com/tpluscode/fun-ddr/commit/0765177))
+* repository stores empty graph because it spreads wrong object ([7ecf8e4](https://github.com/tpluscode/fun-ddr/commit/7ecf8e4))
+
+
+
+
+
+## [0.1.3](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.1.2...@tpluscode/fun-ddr-sparql-graph-repository@0.1.3) (2019-10-02)
+
+
+### Bug Fixes
+
+* make aggregate root async to wait for modifications to finish ([7634cc9](https://github.com/tpluscode/fun-ddr/commit/7634cc9))
+
+
+
+
+
 ## [0.1.2](https://github.com/tpluscode/fun-ddr/compare/@tpluscode/fun-ddr-sparql-graph-repository@0.1.1...@tpluscode/fun-ddr-sparql-graph-repository@0.1.2) (2019-10-01)